+++ change-log.md
@@ -1,9 +1,10 @@
-## Change Log - Charlie
-Created At: 18/10/2018
+# Change Log
+ - Created At: 18/10/2018
+ - Last Updated At: 22/10/2018
 
-## Update Database
+# Update Database
 
-
+## 18/10/2018
 **Create new record in SavvyLoaderJob**
 
 > *Warning: Must check source_folder, success_folder and fail_folder before run code*
@@ -117,3 +118,23 @@ def sql_mercari_merge_statement(dest_table, all_fields, key_fields):
 
 ```
 
+---
+
+## 22/10/2018
+
+Change name table Environmental_Price_MercariClosingPrices to Environmental_Prices_MercariClosingPrices
+
+> Using Procedures
+``` 
+EXEC sp_rename 'Environmental_Prices_MercariClosingPrices', 'Environmental_Price_MercariClosingPrices'
+```
+
+update code match a day in filename
+
+``` 
+...
+    date_match = search(r'\d{4}-\d{2}-\d{2}', fname)
+    if date_match.group():
+        df['added_dttm'] = date_match.group();
+...        
+```